re-write of merge_strategy="merge"
previously, this was failing on corner cases where some duplicate IDs
should be merged but some should not.  Now with the duplicates table in
the db and this re-write, it's possible to check all previously-imported
features whose id_spec evaluated to the same as the current feature, and
see which (if any) of those can be merged.  Way more robust.