Handle a bug in NCBI gtf (#172)
NCBI gtf sometimes has an empty transcript_id field, which causes an
IndexError when creating a db because gffutils sees that there is a
transcript_id attribute present and then tries to use it depsite it
being blank. This commit adds a second check to make sure that it is not
blank before accessing it to avoid this error.